python/requirements-web.txt
Flask==3.1.2
requests==2.32.5
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0

